Need to update uploader.json in first phase
Do it in this phase because it's unprivileged.  I also needed to modify
what is returned from InsightsClient.update() so that I could tell if it
was invoked or not due to the whole client_startup() thing.

Also threw in a bug fix in InsightsClient.update() where the verify()
return value wasn't being read properly.