always run a new archive collection (#406)
The original problem that cached collections solved was to enable uploading to be attempted multiple times without having to collect again.  Turns out that solution was too complex for now at least, so we're just going back to using temp files and passing the path to the upload phase multiple times if it fails.