[wip] Refactor connection_rules.InsightsUploadConf.get_conf (#1230)
* Refactor connection_rules.InsightsConfig.get_conf

The get_conf method was bulky with many conditionals. Reduce its
congnitive complexity by splitting the method into smaller ones. Use
separate methods for different configuration sources.

InsightsConfig interface changed, existing behavior remains the same.

* Create regression tests

Created tests suites for the changed pieces of code. The tests are not
DRY and do not test the actual introduced methods. The reason for
that is that they test that the behavior didn’t change with the
refactoring. It is not intended to leave them like this. They are left
in this commit for reference.

* Test separate get_conf_* methods

Create separate tests for InsightsUploadConf.get_conf_*
methods. Modify existing collect function tests to trust those methods
and not to tests their internal working.

* Add back some regression tests

The newly introduced unit tests do not check whether the functionality
remained the same after the code changes. Added some regression tests
that check the actual external behavior of the changed pieces of
code.

These tests do not check the internal workings of the methods
more than necessary to actually prove they are doing what they
should. Generally the main stitch points are tested: errors, signature
validations, return values.