Add tags.conf (#2289)
* Add get_tags utility function

Signed-off-by: Link Dupont <link@sub-pop.net>

* Collect and add tags to archive

Signed-off-by: Link Dupont <link@sub-pop.net>

* Try importing constants from within package...

...otherwise import it as a relative module. This fixes running
config.py as a __main__ script.

Signed-off-by: Link Dupont <link@sub-pop.net>

* Add write_tags function to write out tags file

Signed-off-by: Link Dupont <link@sub-pop.net>

* Write "group" config if passed as arg

Signed-off-by: Link Dupont <link@sub-pop.net>

* Add test_write_tags

Signed-off-by: Link Dupont <link@sub-pop.net>

* Wrap tags inside "insights-client" namespace

Signed-off-by: Link Dupont <link@sub-pop.net>

* Use python 2/3 compatible error handling

Signed-off-by: Link Dupont <link@sub-pop.net>

* Create tags dict if get_tags returns None

Signed-off-by: Link Dupont <link@sub-pop.net>

* Set `default_flow_style` to `False`

So we don't use a flow style that looks identical to JSON...

https://pyyaml.org/wiki/PyYAMLDocumentation#Dictionarieswithoutnestedcollectionsarenotdumpedcorrectly
Signed-off-by: Link Dupont <link@sub-pop.net>

* Encode tags dict as a list of dicts

This encoding process loops over each key/value pair in tags and
performs one of three actions:
1) If the value is a list, the encoding function is called recursively
on each element in the list. The key is repeated for each element.
The results are collected and chained into a list, which is then
returned.
2) If the value is a dict, the encoding function is called recursively
on each key/value pair in the dict. The key is first prefixed with the
"parent" key plus a ":" (colon) as a scope delimiter. The results are
collected and chained into a list, which is then returned.
3) All over values return a dictionary with two entries: "key" has a
value of the current entry key. "value" has a value of the current
entry value. An additional "namespace" entry is added, with a
value pulled from `constants.app_name`.

Signed-off-by: Link Dupont <link@sub-pop.net>