RHCLOUD-8745 Fix warning when skipping a glob (#2782)
* Fix warning when skipping a glob

When a file matched by a glob pattern gets skipped, a message with a
whole dictionary is printed instead of just the file name. Fixed. The
warning is now consistent with other skipped matches.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Test skipped glob warning

Added a test for the warning emitted when a file matched by a glob is
skipped. Originally a dictionary was output instead of the file name.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Test other skip warnings

The skipped file warning was one of a few more that use the same
logic. Added tests for the sibling calls so they are all covered.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

Co-authored-by: Jeremy Crafts <gravitypriest@users.noreply.github.com>